Want simple performance tool
This tool allows us to pump IO full speed to supported storage URI's.
The intent of this is to be able to test performance without the need
for setting up any targets and, subsequently, benchmark tools.

Its no where near as sophisticated as fio, but thats also not the
intent. The intent is to get good numbers but also, to assist in
testing/developing the handling of IO errors and rebuild

+++ mayastor/src/core/handle.rs
@@ -75,7 +75,7 @@ impl BdevHandle {
 
     /// Allocate memory from the memory pool (the mem is zeroed out)
     /// with given size and proper alignment for the bdev.
-    pub fn dma_malloc(&self, size: usize) -> Result<DmaBuf, DmaError> {
+    pub fn dma_malloc(&self, size: u64) -> Result<DmaBuf, DmaError> {
         DmaBuf::new(size, self.desc.get_bdev().alignment())
     }
 
@@ -141,7 +141,7 @@ impl BdevHandle {
         &self,
         offset: u64,
         buffer: &mut DmaBuf,
-    ) -> Result<usize, CoreError> {
+    ) -> Result<u64, CoreError> {
         let (s, r) = oneshot::channel::<bool>();
         let errno = unsafe {
             spdk_bdev_read(
